requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
reportlab==4.4.4
gunicorn==20.1.0
